refactor(iota-core): split authority_client into per-service modules
