+++ crates/iota-core/src/authority_server/validator_peer.rs
@@ -1,4 +1,4 @@
-// Modifications Copyright (c) 2026 IOTA Stiftung
+// Copyright (c) 2026 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use iota_network::{
