+++ crates/iota-e2e-tests/tests/zklogin_tests.rs
@@ -5,7 +5,7 @@
 
 use std::net::SocketAddr;
 
-use iota_core::authority_client::AuthorityAPI;
+use iota_core::authority_client::validator::ValidatorAPI;
 use iota_macros::sim_test;
 use iota_sdk_types::crypto::{Intent, IntentMessage};
 use iota_test_transaction_builder::TestTransactionBuilder;
